requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.5.0
Pillow>=10.0.0
